+++ lib/windshaft/renderers/pg-mvt/renderer.js
