+++ lib/windshaft/renderers/pg-mvt/sql.js
@@ -1,4 +1,4 @@
-var debug = require('debug')('windshaft:pg_mvt');
+var debug = require('debug')('windshaft:pg-mvt');
 
 // TODO copied from torque, extract somewhere
