+++ lib/windshaft/renderers/renderer_factory.js
@@ -3,7 +3,7 @@ var BlendRenderer = require('./blend');
 var TorqueRenderer = require('./torque');
 var MapnikRenderer = require('./mapnik');
 var PlainRenderer = require('./plain');
-var PgMvtRenderer = require('./pg_mvt');
+var PgMvtRenderer = require('./pg-mvt');
 var layersFilter = require('../utils/layer_filter');
 
 var step = require('step');
